Implement navigation between current views using react router
