+++ src/components/layout/Spinner.jsx
@@ -1,18 +1,15 @@
 import React from "react";
-import { useAppContext } from "../../context/app/appContext";
 import spinner from "./spinner.gif";
 import styled from "styled-components";
 
 const StyledSpinner = styled.img`
   width: 50px;
   margin: auto;
-  padding-top: 200px;
+  padding: 50px 0 100px;
   display: block;
 `;
 const Spinner = () => {
-  const { loading } = useAppContext();
-
-  return loading && <StyledSpinner src={spinner} alt="Loading..." />;
+  return <StyledSpinner src={spinner} alt="Loading..." />;
 };
 
 export default Spinner;
